+++ src/main/java/frc/robot/intake_assist/IntakeAssistManager.java
@@ -1,4 +1,4 @@
-package frc.robot.note_assist;
+package frc.robot.intake_assist;
 
 import frc.robot.imu.ImuSubsystem;
 import frc.robot.localization.LocalizationSubsystem;
@@ -6,15 +6,15 @@ import frc.robot.swerve.SwerveSubsystem;
 import frc.robot.util.scheduling.LifecycleSubsystem;
 import frc.robot.util.scheduling.SubsystemPriority;
 
-public class NoteAssistManager extends LifecycleSubsystem {
+public class IntakeAssistManager extends LifecycleSubsystem {
   private static final String LIMELIGHT_NAME = "limelight-note";
   private final LocalizationSubsystem localization;
   private final SwerveSubsystem swerve;
   private final ImuSubsystem imu;
 
-  public NoteAssistManager(
+  public IntakeAssistManager(
       LocalizationSubsystem localization, SwerveSubsystem swerve, ImuSubsystem imu) {
-    super(SubsystemPriority.NOTE_ASSIST_MANAGER);
+    super(SubsystemPriority.INTAKE_ASSIST_MANAGER);
 
     this.localization = localization;
     this.swerve = swerve;
